bi_app/backend/requirements.txt
# Django Backend - SIGETI BI
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
pandas==2.1.4
